Split cad.py into microgen.cad package
Replace the single monolithic microgen/cad.py with a microgen.cad subpackage. The functionality has been split across new modules: __init__.py, _install.py, io.py, meshbridge.py, primitives.py, shape.py and topo.py to isolate responsibilities and enable lazy OCP imports (OCP remains an optional dependency). Tests were updated to reflect the package layout (no top-level OCP import). This reorganization improves maintainability and ensures import-time safety when OCP is not installed.
